Change temp type to int16

+++ app/src/hw_module.h
@@ -1,7 +1,7 @@
 #ifndef hw_module_h
 #define hw_module_h
 
-int32_t hpi_hw_read_temp(void);
+int16_t hpi_hw_read_temp(void);
 
 void send_usb_cdc(const char *buf, size_t len);
